+++ modules/atomic-opt-in/assets/js/components/admin-tab-content.js
