+++ modules/atomic-opt-in/assets/js/components/admin-tab-img-placeholders.js
